strands-agent/requirements.txt
# Core dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic>=2.8.0,<3.0.0
python-multipart==0.0.9

# AWS Strands Agents SDK
strands-agents==0.1.0
strands-agents-tools==0.1.0

# LLM SDKs
anthropic==0.31.0
boto3==1.34.131  # For AWS Bedrock

# Database
asyncpg==0.29.0
sqlalchemy==2.0.31
alembic==1.13.2

# Vector database
qdrant-client==1.10.1

# MCP integration
httpx==0.27.0
aiofiles==24.1.0
mcp>=1.8.0

# Utilities
python-dotenv==1.0.1
loguru==0.7.2
tenacity>=9.1.2
python-jose==3.3.0
pyyaml==6.0.1

# Development
pytest==8.2.2
pytest-asyncio==0.23.7
black==24.4.2
ruff==0.5.0